requirements.txt
aiofiles==0.7.0
alembic==1.7.4
anyio==3.3.3
asgiref==3.4.1
async==0.6.2
asyncpg==0.24.0
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.2
httpx==0.21.1
idna==3.2
importlab==0.6.1
iniconfig==1.1.1
libcst==0.3.21
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
networkx==2.6.3
ninja==1.10.2.2
orjson==3.6.4
packaging==21.2
passlib==1.7.4
password==0.2
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.11.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.16.0
python-dotenv==0.19.1
python-jose==3.3.0
python-multipart==0.0.5
pytype==2021.11.12
PyYAML==6.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.25
starlette==0.16.0
tabulate==0.8.9
toml==0.10.2
typed-ast==1.5.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
uvicorn==0.15.0
